client/requirements.txt
dnspython
fastapi
uvicorn
PyYAML
